Add google checks for clang-tidy and fix resulting warnings

+++ src/normal3.cpp
@@ -76,14 +76,14 @@ math::Normal3& math::Normal3::operator*=(float Scalar)
 math::Normal3 math::Normal3::operator/(float Scalar) const
 {
     assert(Scalar != 0);
-    const float Rec = (float)1 / Scalar;
+    const float Rec = 1.0f / Scalar;
     return {X * Rec, Y * Rec, Z * Rec};
 }
 
 math::Normal3& math::Normal3::operator/=(float Scalar)
 {
     assert(Scalar != 0);
-    const float Rec = (float)1 / Scalar;
+    const float Rec = 1.0f / Scalar;
     X *= Rec;
     Y *= Rec;
     Z *= Rec;
