+++ src/vector2.cpp
@@ -69,14 +69,14 @@ math::Vector2& math::Vector2::operator*=(float Scalar)
 math::Vector2 math::Vector2::operator/(float Scalar) const
 {
     assert(Scalar != 0);
-    const float Rec = (float)1 / Scalar;
+    const float Rec = 1.0f / Scalar;
     return {X * Rec, Y * Rec};
 }
 
 math::Vector2& math::Vector2::operator/=(float Scalar)
 {
     assert(Scalar != 0);
-    const float Rec = (float)1 / Scalar;
+    const float Rec = 1.0f / Scalar;
     X *= Rec;
     Y *= Rec;
     return *this;
